refactor: merge AuthTokenDTOs into one file
